Keep ordinal bookkeeping out of the render strategies
The ordinal names an ovstage publication, and the ovstage path always
renders synchronously, so it now steps the renderer directly. The
strategy interface serves the legacy path and loses the ordinal
parameter. The ordinal-recovery test moves to the renderer contract
tests, and the strategy-level ordinal propagation tests are removed
with the parameter.
