+++ source/isaaclab_ov/test/test_ovrtx_renderer_contract.py
@@ -697,3 +697,48 @@ def test_ovrtx_close_is_idempotent():
     renderer.close()
 
     assert events == []
+
+
+def test_ovstage_ordinal_advances_when_consumption_fails():
+    """A failed product consumption must not leave the ordinal at the barred write floor.
+
+    ``advance_write_floor`` runs before the step and bars writes at the current ordinal. A stale
+    ordinal would reject every later scene write from a caller that catches the render error and
+    keeps stepping.
+    """
+
+    class _Completion:
+        def wait(self) -> None:
+            return None
+
+    class _Stage:
+        def advance_write_floor(self, **_kwargs) -> _Completion:
+            return _Completion()
+
+    class _StepRecorder:
+        def __init__(self):
+            self.ordinals: list[int] = []
+
+        def step(self, render_products, delta_time, *, ordinal):
+            self.ordinals.append(ordinal)
+            return {}
+
+    renderer = _make_ovrtx_renderer_without_backend()
+    renderer._initialized_scene = True
+    renderer._use_ovstage = True
+    renderer._renderer = _StepRecorder()
+    renderer._render_product_paths = ["/product"]
+    renderer._visual_material_writer_ref = None
+    renderer._stage = _Stage()
+    renderer._current_ordinal = 7
+
+    def _failing_consume(render_data, products):
+        raise RuntimeError("extraction failed")
+
+    renderer._consume_products = _failing_consume
+
+    with pytest.raises(RuntimeError, match="extraction failed"):
+        renderer._render_ovstage(object())
+
+    assert renderer._renderer.ordinals == [7]
+    assert renderer._current_ordinal == 8
